Properly implements wing sweep.
Although most of the structure was in for handling sweep, there were
bugs in how the positions of the control and corner points of the
vortices and the section areas were being calculated in the case of
sweep. Tests were implemented to test for correct function in these
areas and the bugs were fixed.