Adds flag for linear spacing of vortices.
The default is for the LLGrid to use cosine spacing for the
horseshoe vortices. This allows for linear spacing to be used if
desired. Although there aren't any existing lifting-line results to
compare the linear spacing to, a grid convergence study was
performed and the linearly spaced cases converged to the same values
as the cosine spaced cases as more nodes were added.